pyproject.toml
[tool.poetry]
name = "modulbankintegration"
version = "0.1.0"
description = ""
authors = ["Pavel2232 <pavelmazikinvladimirovich@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.9"
redis = "^5.0.1"
celery = "^5.3.6"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
django-celery-beat = "^2.5.0"
requests = "^2.31.0"
pydantic = "^2.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
